INS-277: move code parts in separate files
