+++ api/query_type.go
@@ -1,0 +1,45 @@
+/*
+ *    Copyright 2018 INS Ecosystem
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package main
+
+type QueryType int
+
+const (
+	UNDEFINED QueryType = iota
+	CreateMember
+	DumpUserInfo
+	GetBalance
+	SendMoney
+	DumpAllUsers
+)
+
+func QTypeFromString(strQType string) QueryType {
+	switch strQType {
+	case "create_member":
+		return CreateMember
+	case "dump_user_info":
+		return DumpUserInfo
+	case "get_balance":
+		return GetBalance
+	case "send_money":
+		return SendMoney
+	case "dump_all_user":
+		return DumpAllUsers
+	}
+
+	return UNDEFINED
+}
